+++ opennmt/CHANGELOG.md
@@ -3,7 +3,60 @@
 
 
 ## [Unreleased]
+
+## [1.1.1](https://github.com/OpenNMT/OpenNMT-py/tree/1.1.1) (2020-03-20)
 ### Fixes and improvements
+* Fix backcompatibility when no 'corpus_id' field (c313c28)
+
+## [1.1.0](https://github.com/OpenNMT/OpenNMT-py/tree/1.1.0) (2020-03-19)
+### New features
+* Support CTranslate2 models in REST server (91d5d57)
+* Extend support for custom preprocessing/postprocessing function in REST server by using return dictionaries (d14613d, 9619ac3, 92a7ba5)
+* Experimental: BART-like source noising (5940dcf)
+
+### Fixes and improvements
+* Add options to CTranslate2 release (e442f3f)
+* Fix dataset shard order (458fc48)
+* Rotate only the server logs, not training (189583a)
+* Fix alignment error with empty prediction (91287eb)
+
+## [1.0.2](https://github.com/OpenNMT/OpenNMT-py/tree/1.0.2) (2020-03-05)
+### Fixes and improvements
+* Enable CTranslate2 conversion of Transformers with relative position (db11135)
+* Adapt `-replace_unk` to use with learned alignments if they exist (7625b53)
+
+## [1.0.1](https://github.com/OpenNMT/OpenNMT-py/tree/1.0.1) (2020-02-17)
+### Fixes and improvements
+* Ctranslate2 conversion handled in release script (1b50e0c)
+* Use `attention_dropout` properly in MHA (f5c9cd4)
+* Update apex FP16_Optimizer path (d3e2268)
+* Some REST server optimizations
+* Fix and add some docs
+
+## [1.0.0](https://github.com/OpenNMT/OpenNMT-py/tree/1.0.0) (2019-10-01)
+### New features
+* Implementation of "Jointly Learning to Align & Translate with Transformer" (@Zenglinxiao)
+
+### Fixes and improvements
+* Add nbest support to REST server (@Zenglinxiao)
+* Merge greedy and beam search codepaths (@Zenglinxiao)
+* Fix "block ngram repeats" (@KaijuML, @pltrdy)
+* Small fixes, some more docs
+
+## [1.0.0.rc2](https://github.com/OpenNMT/OpenNMT-py/tree/1.0.0.rc1) (2019-10-01)
+* Fix Apex / FP16 training (Apex new API is buggy)
+* Multithread preprocessing way faster (Thanks @francoishernandez)
+* Pip Installation v1.0.0.rc1 (thanks @pltrdy)
+
+## [0.9.2](https://github.com/OpenNMT/OpenNMT-py/tree/0.9.2) (2019-09-04)
+* Switch to Pytorch 1.2
+* Pre/post processing on the translation server
+* option to remove the FFN layer in AAN + AAN optimization (faster)
+* Coverage loss (per Abisee paper 2017) implementation
+* Video Captioning task: Thanks Dylan Flaute!
+* Token batch at inference
+* Small fixes and add-ons
+
 
 ## [0.9.1](https://github.com/OpenNMT/OpenNMT-py/tree/0.9.1) (2019-06-13)
 * New mechanism for MultiGPU training "1 batch producer / multi batch consumers"